requirements.txt
Flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.0.33
bcrypt==4.0.1
Flask-Mail==0.9.1
itsdangerous==2.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7
